Generalizing generators for both metrics test suites
